actionsheet [nfc]: Move code to `src/action-sheet/index.js`.
This is a better place because the file contains all action sheets
and not just `MessageActionSheet`.
