+++ src/streams/TopicItem.js
@@ -8,8 +8,8 @@ import invariant from 'invariant';
 
 import styles, { BRAND_COLOR, createStyleSheet } from '../styles';
 import { RawLabel, Touchable, UnreadCount } from '../common';
-import { showTopicActionSheet } from '../message/messageActionSheet';
-import type { ShowActionSheetWithOptions } from '../message/messageActionSheet';
+import { showTopicActionSheet } from '../action-sheets';
+import type { ShowActionSheetWithOptions } from '../action-sheets';
 import { TranslationContext } from '../boot/TranslationProvider';
 import { useDispatch, useSelector } from '../react-redux';
 import {
