+++ src/title/TitleStream.js
@@ -21,8 +21,8 @@ import {
   getOwnUser,
   getStreamInNarrow,
 } from '../selectors';
-import { showTopicActionSheet } from '../message/messageActionSheet';
-import type { ShowActionSheetWithOptions } from '../message/messageActionSheet';
+import { showTopicActionSheet } from '../action-sheets';
+import type { ShowActionSheetWithOptions } from '../action-sheets';
 import { getUnread } from '../unread/unreadModel';
 
 type Props = $ReadOnly<{|
